Examinations Razor Pages

+++ LabApp/Dtos/ExaminationDto.cs
@@ -18,7 +18,7 @@
         public string Unit { get; set; }
         public string LowerRange { get; set; }
         public string UpperRange { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
         public int OrderId { get; set; }
         public int DeviceId { get; set; }
@@ -31,7 +31,7 @@
         public string Unit { get; set; }
         public string LowerRange { get; set; }
         public string UpperRange { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
     }
 
 }
